server/requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
pyjwt
python-decouple
twilio